Added a compression utility for releases
The compressor script creates gzip and zip files for the specified build type ('release,' 'package,' or 'all'). Compressed files are saved to the 'dist' directory with release names based on the project name and the current version. A file with the same name already exists will iterate the sub-version number (e.g., 0.5.3.1, 0.5.3.2, and so on) to make the file name unique.

For the release build type, it will compress the specified files (manifest.json, main.js, and style.css) and create files with the prefix release.

The 'package' build will compress all files, excluding anything listed in the .gitignore file.